openharmony-schedule - 30

+++ openharmony/arkcompiler_ets_runtime/ecmascript/mem/mem.h
@@ -37,8 +37,8 @@ enum class MemAlignmentLog2 : uint8_t {
     MEM_ALIGN_REGION_LOG2 = 4,
 };
 
-static constexpr size_t INITIAL_REGULAR_OBJECT_CAPACITY = 1024_MB;
-static constexpr size_t INITIAL_HUGE_OBJECT_CAPACITY = 1024_MB;
+static constexpr uint64_t INITIAL_REGULAR_OBJECT_CAPACITY = 1024_MB;
+static constexpr uint64_t INITIAL_HUGE_OBJECT_CAPACITY = 1024_MB;
 static constexpr size_t INCREMENT_HUGE_OBJECT_CAPACITY = 128_MB;
 static constexpr size_t LARGE_POOL_SIZE = 480_MB;
 static constexpr size_t MEDIUM_POOL_SIZE = 256_MB;
